+++ src/components/FormElements/Input/Input.tsx
@@ -1,7 +1,9 @@
-import React, { forwardRef, Ref, ForwardRefRenderFunction } from "react";
+import React, { forwardRef, ForwardRefRenderFunction, useImperativeHandle, useRef } from "react";
 import classNames from "classnames";
 import { SerializedStyles } from "@emotion/react";
 import Label from "../Label/Label";
+import { InfoCircledSVG, CloseSVG } from "../../../icons/index";
+import Tooltip from "../../Tooltip/Tooltip";
 import { inputContainer } from "./styles";
 import { ExtendableProps, IconType } from "types/common";
 
@@ -15,11 +17,14 @@ export type InputProps = ExtendableProps<
     size?: InputSize;
     iconBefore?: IconType;
     iconAfter?: IconType;
-    ref?: Ref<HTMLInputElement>;
     label?: string;
     inline?: boolean;
     containerAttrs?: React.HTMLAttributes<HTMLDivElement>;
     css?: SerializedStyles;
+    tooltipContent?: string;
+    showVerticalLine?: boolean;
+    isClearable?: boolean;
+    onClear?: () => void;
   }
 >;
 
@@ -33,10 +38,17 @@ const Input: ForwardRefRenderFunction<HTMLInputElement, InputProps> = (
     inline = false,
     id,
     containerAttrs,
+    tooltipContent = "",
+    value,
+    isClearable = false,
+    showVerticalLine = true,
+    onClear,
     ...rest
   },
   forwardedRef,
 ) => {
+  const internalRef = useRef<HTMLInputElement>(null);
+
   const IconBefore = iconBefore;
   const IconAfter = iconAfter;
   const hasLabel = Boolean(label);
@@ -50,24 +62,62 @@ const Input: ForwardRefRenderFunction<HTMLInputElement, InputProps> = (
   });
   const iconHeight = size === "sm" ? 28 : 32;
 
+  useImperativeHandle<HTMLInputElement | null, HTMLInputElement | null>(
+    forwardedRef,
+    () => internalRef.current,
+  );
+
+  const setFocus = () => {
+    internalRef.current?.focus();
+  };
+
+  const handleClear = () => {
+    if (onClear) onClear();
+    setFocus();
+  };
+
   return (
     <div
-      css={(theme): SerializedStyles => inputContainer(theme, { size })}
+      css={(theme): SerializedStyles =>
+        inputContainer(theme, {
+          size,
+          hasIconAfter: Boolean(iconAfter),
+          isClearable,
+          showVerticalLine,
+        })
+      }
       className={containerClasses}
       {...containerAttrs}
     >
-      {hasLabel && <Label htmlFor={id}>{label}</Label>}
+      {hasLabel && (
+        <div className="label-container">
+          <Label htmlFor={id}>{label}</Label>
+          {tooltipContent?.length > 0 && (
+            <Tooltip content={tooltipContent}>
+              <InfoCircledSVG height={20} />
+            </Tooltip>
+          )}
+        </div>
+      )}
       <div className="input-wrapper">
         {IconBefore && (
           <span className="prefix-icon" data-testid="input-icon-before">
             <IconBefore height={iconHeight} />
           </span>
         )}
-        <input ref={forwardedRef} id={id} {...rest} />
+        <input value={value} ref={internalRef} id={id} {...rest} />
         {IconAfter && (
-          <span className="suffix-icon" data-testid="input-icon-after">
-            <IconAfter height={iconHeight} />
-          </span>
+          <>
+            {showVerticalLine && <div className="vertical-line" />}
+            <span className="suffix-icon" data-testid="input-icon-after" onClick={setFocus}>
+              <IconAfter height={iconHeight} />
+            </span>
+          </>
+        )}
+        {isClearable && value && (
+          <div className="close-icon" onClick={handleClear}>
+            <CloseSVG height={16} />
+          </div>
         )}
       </div>
     </div>