+++ src/components/Pagination/Pagination.tsx
@@ -5,8 +5,7 @@ import Button from "../Button/Button";
 import { IconChevronDownSVG, RightArrowSVG } from "../../icons/";
 import Text from "../Text/Text";
 import { container } from "./styles";
-import { PaginationProps, RowItem } from "./types";
-import { usePagination } from "./usePagination";
+import { PaginationProps } from "./types";
 import useClickOutside from "./hooks";
 
 const SelectedOptionClasses = (isSelected: boolean): string =>
@@ -15,22 +14,22 @@ const SelectedOptionClasses = (isSelected: boolean): string =>
   });
 
 const Pagination: FC<PaginationProps> = ({
-  current,
-  list,
+  page,
+  pageSize,
   totalPages,
-  size,
+  rowsPerPageOptions,
   selectionText,
-  handlePaginationSizeChanged,
-  handlePaginationNumberChanged,
   dir = "ltr",
+  onPageChange,
+  onPageSizeChange,
   ...rest
 }) => {
-  const [isListOpen, setIsListOpen] = useState(false);
   const wrapperRef = useRef<HTMLDivElement | null>(null);
-  const [listItemSelected, handleListItemSelected] = useState<number>(size);
-  const paginationRange = usePagination(current, totalPages);
-  const hasItems = paginationRange.length > 0;
-
+  const [isListOpen, setIsListOpen] = useState(false);
+  const [selectedListItem, setSelectedListItem] = useState<number>(pageSize);
+  const isRtl = dir === "rtl";
+  const isPrevBtnDisabled = page === 1;
+  const isNextBtnDisabled = page === totalPages || totalPages === 0;
   useClickOutside(wrapperRef, () => setIsListOpen(false));
 
   const toggleList = (): void => {
@@ -40,91 +39,74 @@ const Pagination: FC<PaginationProps> = ({
 
   const handleListItemSelect = (item: number): void => {
     if (isListOpen) {
-      handlePaginationSizeChanged(item);
-      handleListItemSelected(item);
+      onPageSizeChange(item);
+      setSelectedListItem(item);
       setIsListOpen(false);
     }
   };
 
-  const mapOptions = (items: RowItem[]): JSX.Element[] => {
-    return items
-      .slice()
-      .reverse()
-      .map((item) => {
-        const isSelected = item.id === listItemSelected;
-
-        return (
-          <li
-            data-testid="pagination-page"
-            key={"item" + item}
-            onClick={(): void => handleListItemSelect(item.id)}
-          >
-            <Text fontSize={"sm"} className={SelectedOptionClasses(isSelected)}>
-              {item.value}
-            </Text>
-          </li>
-        );
-      });
-  };
+  const options = rowsPerPageOptions.slice().reverse();
 
   return (
-    <>
-      {hasItems && (
-        <div css={(theme): SerializedStyles => container(theme, { isOpen: isListOpen })} {...rest}>
-          <Button
-            className="previous-page-btn"
-            data-testid="previous-page-btn"
-            name="Previous page"
-            onClick={(): void => handlePaginationNumberChanged(current - 1)}
-            variant="ghost"
-            noGutters
-            disabled={current === 1}
-          >
-            {dir === "rtl" ? (
-              <IconChevronDownSVG className="left" height={32} />
-            ) : (
-              <IconChevronDownSVG className="right" height={32} />
-            )}
-          </Button>
+    <div css={(theme): SerializedStyles => container(theme, { isOpen: isListOpen })} {...rest}>
+      <Button
+        className="previous-page-btn"
+        data-testid="previous-page-btn"
+        name="Previous page"
+        onClick={(): void => onPageChange(page - 1)}
+        variant="ghost"
+        noGutters
+        disabled={isPrevBtnDisabled}
+      >
+        <IconChevronDownSVG className={isRtl ? "rotate-left" : "rotate-right"} height={32} />
+      </Button>
 
-          <div className="pagination-options">
-            <div className="dropdown" ref={wrapperRef}>
-              <div className="dropdown-button" onClick={toggleList}>
-                <Button iconAfter={RightArrowSVG} variant="ghost">
-                  <Text fontSize="sm">{selectionText}</Text>
-                </Button>
-              </div>
-
-              {isListOpen && list.length > 0 && (
-                <div className="open-list-container">
-                  <div className="dropdown-wrapper">
-                    <ul role="list" className="dropdown-list">
-                      {mapOptions(list)}
-                    </ul>
-                  </div>
-                </div>
-              )}
-            </div>
+      <div className="pagination-options">
+        <div className="dropdown" ref={wrapperRef}>
+          <div className="dropdown-button" onClick={toggleList}>
+            <Button iconAfter={RightArrowSVG} variant="ghost">
+              <Text fontSize="sm">{selectionText}</Text>
+            </Button>
           </div>
 
-          <Button
-            className="next-page-btn"
-            data-testid="next-page-btn"
-            name="Next page"
-            onClick={(): void => handlePaginationNumberChanged(current + 1)}
-            variant="ghost"
-            noGutters
-            disabled={current === paginationRange.length}
-          >
-            {dir === "rtl" ? (
-              <IconChevronDownSVG className="right" height={32} />
-            ) : (
-              <IconChevronDownSVG className="left" height={32} />
-            )}
-          </Button>
+          {isListOpen && rowsPerPageOptions.length > 0 && (
+            <div className="open-list-container">
+              <div className="dropdown-wrapper">
+                <ul role="list" className="dropdown-list">
+                  {options.map(({ value, label }) => {
+                    const isSelected = value === selectedListItem;
+
+                    return (
+                      <li
+                        data-testid="pagination-page"
+                        key={`item-${value}`}
+                        onClick={(): void => handleListItemSelect(value)}
+                      >
+                        <Text fontSize="sm" className={SelectedOptionClasses(isSelected)}>
+                          {label}
+                        </Text>
+                      </li>
+                    );
+                  })}
+                </ul>
+              </div>
+            </div>
+          )}
         </div>
-      )}
-    </>
+      </div>
+
+      <Button
+        className="next-page-btn"
+        data-testid="next-page-btn"
+        name="Next page"
+        onClick={(): void => onPageChange(page + 1)}
+        variant="ghost"
+        noGutters
+        disabled={isNextBtnDisabled}
+      >
+        <IconChevronDownSVG className={isRtl ? "rotate-right" : "rotate-left"} height={32} />
+      </Button>
+    </div>
   );
 };