+++ src/components/Table/Table.stories.tsx
@@ -23,22 +23,51 @@ export default {
   },
   args: {
     columns: [
-      { accessor: "id", cell: "Code", classNames: ["id"] },
-      { accessor: "description", cell: "Description", classNames: ["description"] },
-      { accessor: "name", cell: "Name", classNames: ["name"] },
-      { accessor: "category", cell: "Category", classNames: ["category"] },
+      { accessor: "id", cell: "Code", classNames: ["id"], sortOrder: "asc" },
+      {
+        accessor: "description",
+        cell: "Description",
+        classNames: ["description"],
+        sortOrder: "asc",
+        sortableHeader: false,
+      },
+      { accessor: "name", cell: "Name", classNames: ["name"], sortOrder: "asc" },
+      { accessor: "category", cell: "Category", classNames: ["category"], sortOrder: "asc" },
+      { accessor: "date", cell: "Date", classNames: ["date"], sortOrder: "desc" },
     ],
     rows: [
       {
         id: 271,
-        description: "Lorem Ipsum is simply dummy text of the printing and typesetting industry. ",
+        description: "Test",
         name: "Test",
         category: "Test",
         code: "Test",
+        date: "27/05/1997",
       },
-      { id: 272, description: "Test", name: "Test", category: "Test", code: "Test" },
-      { id: 273, description: "Test", name: "Test", category: "Test", code: "Test" },
-      { id: 274, description: "Test", name: "Test", category: "Test", code: "Test" },
+      {
+        id: 272,
+        description: "Test",
+        name: "Test",
+        category: "Test",
+        code: "Test",
+        date: "27/05/1997",
+      },
+      {
+        id: 273,
+        description: "Test",
+        name: "Test",
+        category: "Test",
+        code: "Test",
+        date: "27/05/1997",
+      },
+      {
+        id: 274,
+        description: "Test",
+        name: "Test",
+        category: "Test",
+        code: "Test",
+        date: "27/05/1997",
+      },
     ],
     emptyState: {
       title: "No results found with these criteria",
@@ -59,8 +88,7 @@ export const Default = Template.bind({});
 export const WithSorting = Template.bind({});
 
 WithSorting.args = {
-  sortable: true,
-  sorting: { column: "id", isDescending: true },
+  sorting: { column: "id", isDescending: false },
   onSortingChanged: (sorting: Sorting) => console.log(sorting),
 };
 
@@ -72,12 +100,63 @@ WithRowSelection.args = {
   onRowClick: (row: Row) => console.log(row),
 };
 
+export const WithOverflowColumns = Template.bind({});
+
+WithOverflowColumns.args = {
+  columns: [
+    { accessor: "id", cell: "Code", classNames: ["id"], sortOrder: "asc", headerWidth: 100 },
+    {
+      accessor: "description",
+      cell: "Description",
+      classNames: ["description"],
+      maxWidth: 100,
+      sortOrder: "asc",
+    },
+    {
+      accessor: "name",
+      cell: "Name",
+      classNames: ["name"],
+      maxWidth: 100,
+      sortOrder: "asc",
+    },
+    { accessor: "category", cell: "Category", classNames: ["category"], sortOrder: "asc" },
+  ],
+  rows: [
+    {
+      id: 271,
+      description: <div className="has-overflow">Testing Testing Testing</div>,
+      name: "Test",
+      category: "Test",
+      code: "Test",
+    },
+    { id: 272, description: "Test", name: "Test", category: "Test", code: "Test" },
+    {
+      id: 273,
+      description: "Test",
+      name: (
+        <div className="has-overflow">
+          Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum is
+          simply dummy text of the printing and typesetting industry. Lorem Ipsum is simply dummy
+          text of the printing and typesetting industry. Lorem Ipsum is simply dummy text of the
+          printing and typesetting industry. Lorem Ipsum is simply dummy text of the printing and
+          typesetting industry. Lorem Ipsum is simply dummy text of the printing and typesetting
+          industry.
+        </div>
+      ),
+      category: "Test",
+      code: "Test",
+    },
+    { id: 274, description: "Test", name: "Test", category: "Test", code: "Test" },
+  ],
+};
+
 export const WithoutData = Template.bind({});
 
 WithoutData.args = {
   rows: [],
   emptyState: {
     ...emptyState,
+    hideInfo: true,
     footer: (
       <div className="body">
         <Text fontSize="lg">{emptyState.title}</Text>