+++ src/components/Table/components/Cell.tsx
@@ -1,14 +1,45 @@
-import React, { FC, HTMLAttributes } from "react";
+import React, { FC, HTMLAttributes, useEffect, useLayoutEffect, useRef, useState } from "react";
+import Tooltip from "../../Tooltip/Tooltip";
 
 export type CellProps = HTMLAttributes<HTMLTableCellElement> & {
   as?: "td" | "th";
+  windowSize?: number[];
   onClick?: () => void;
 };
 
-const Cell: FC<CellProps> = ({ children, as: Component = "td", onClick, ...rest }) => {
+const Cell: FC<CellProps> = ({
+  children,
+  as: Component = "td",
+  onClick,
+  windowSize,
+  style,
+  ...rest
+}) => {
+  const componentRef = useRef<HTMLTableCellElement | null>(null);
+  const overflowRef = useRef<HTMLElement | null>(null);
+  const [isOverflowActive, setIsOverflowActive] = useState(false);
+
+  useLayoutEffect(() => {
+    if (Component === "td" && componentRef.current) {
+      overflowRef.current = componentRef.current.querySelector(".has-overflow");
+    }
+  }, []);
+
+  useEffect(() => {
+    if (Component === "td") {
+      const el = overflowRef.current;
+
+      if (el) {
+        setIsOverflowActive(el.offsetWidth < el.scrollWidth);
+      }
+    }
+  }, [overflowRef, windowSize]);
+
   return (
-    <Component onClick={onClick} {...rest}>
-      {children}
+    <Component ref={componentRef} style={style} onClick={onClick} {...rest}>
+      <Tooltip content={children} disabled={!isOverflowActive}>
+        <span style={style}>{children}</span>
+      </Tooltip>
     </Component>
   );
 };