+++ src/components/Tabs/Tabs.stories.tsx
@@ -1,40 +1,59 @@
 import React from "react";
 import { Story } from "@storybook/react";
-import { Heading } from "../../";
+import Heading from "../Heading/Heading";
 import Tabs from "./Tabs";
 
 export default {
   title: "Components/Tabs",
 };
 
-const Template: Story = (args) => (
-  <Tabs {...args}>
-    <Tabs.TabPane title="Completed courses">
+const tabs = [
+  {
+    title: "Completed courses",
+    content: (
       <div style={{ padding: "2rem" }}>
         <Heading>Completed courses</Heading>
         <p>Completed courses content!</p>
       </div>
-    </Tabs.TabPane>
-    <Tabs.TabPane title="My super points">
+    ),
+  },
+  {
+    title: "My super points",
+    content: (
       <div style={{ padding: "2rem" }}>
         <Heading>My super points</Heading>
         <p>My super points content!</p>
       </div>
-    </Tabs.TabPane>
-    <Tabs.TabPane title="My superior level">
+    ),
+  },
+  {
+    title: "My superior level",
+    content: (
       <div style={{ padding: "2rem" }}>
         <Heading>My superior level</Heading>
         <p>My superior level content!</p>
       </div>
-    </Tabs.TabPane>
-    <Tabs.TabPane title="I have a lot of certifications">
+    ),
+  },
+  {
+    title: "I have a lot of certifications",
+    content: (
       <div style={{ padding: "2rem" }}>
         <Heading>I have a lot of certifications</Heading>
         <p>I have a lot of certifications content!</p>
       </div>
-    </Tabs.TabPane>
-  </Tabs>
-);
+    ),
+  },
+
+  {
+    title: "I should not be displayed",
+    content: undefined,
+  },
+];
+
+const Template: Story = (args) => {
+  return <Tabs {...args} tabs={tabs} />;
+};
 
 export const Basic = Template.bind({});
 
@@ -69,32 +88,7 @@ OnChangeTab.args = {
 
 const ResponsiveTemplate: Story = (args) => (
   <div style={{ maxWidth: "500px", border: "1px solid red" }}>
-    <Tabs {...args}>
-      <Tabs.TabPane title="Completed courses">
-        <div style={{ padding: "2rem" }}>
-          <Heading>Completed courses</Heading>
-          <p>Completed courses content!</p>
-        </div>
-      </Tabs.TabPane>
-      <Tabs.TabPane title="My super points">
-        <div style={{ padding: "2rem" }}>
-          <Heading>My super points</Heading>
-          <p>My super points content!</p>
-        </div>
-      </Tabs.TabPane>
-      <Tabs.TabPane title="My superior level">
-        <div style={{ padding: "2rem" }}>
-          <Heading>My superior level</Heading>
-          <p>My superior level content!</p>
-        </div>
-      </Tabs.TabPane>
-      <Tabs.TabPane title="I have a lot of certifications">
-        <div style={{ padding: "2rem" }}>
-          <Heading>I have a lot of certifications</Heading>
-          <p>I have a lot of certifications content!</p>
-        </div>
-      </Tabs.TabPane>
-    </Tabs>
+    <Tabs {...args} tabs={tabs} />
   </div>
 );
 
@@ -102,4 +96,11 @@ export const Responsive = ResponsiveTemplate.bind({});
 
 Responsive.args = {
   ...commonProps,
+};
+
+export const withOptionalTabs = Template.bind({});
+
+SelectedTab.args = {
+  ...commonProps,
+  selectedTab: 0,
 };