+++ src/components/Tooltip/Tooltip.tsx
@@ -5,6 +5,7 @@ import { tooltipContainer } from "./styles";
 export type TooltipProps = TippyProps & {
   as?: "div" | "span";
   content: TippyProps["content"];
+  maxWidth?: number;
 };
 
 const Tooltip: FC<TooltipProps> = ({
@@ -12,16 +13,16 @@ const Tooltip: FC<TooltipProps> = ({
   content,
   as = "div",
   placement = "top",
+  maxWidth = 350,
   ...rest
 }) => {
   const Tag = as;
 
   return (
     <Tippy
-      interactive
       placement={placement}
       render={(attrs): ReactNode => (
-        <div className="tooltip" css={tooltipContainer} {...attrs}>
+        <div className="tooltip" css={(theme) => tooltipContainer(maxWidth, theme)} {...attrs}>
           {content}
           <div id="arrow" data-testid="tooltip-arrow" data-popper-arrow />
         </div>